feat!: Moved source into /src folder
