+++ src/commands/nowplaying.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { useQueue } = require('discord-player');
 const { EmbedBuilder, ActionRowBuilder, ButtonBuilder } = require('discord.js');
@@ -5,7 +6,7 @@ const {
     embedColors,
     embedIcons,
     progressBarOptions
-} = require('../config.json');
+} = require(path.resolve('./config.json'));
 
 module.exports = {
     data: new SlashCommandBuilder()
