+++ src/system-commands/guilds.js
@@ -1,10 +1,11 @@
+const path = require('path');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { EmbedBuilder } = require('discord.js');
 const {
     embedColors,
     embedIcons,
     systemServerGuildIds
-} = require('../config.json');
+} = require(path.resolve('./config.json'));
 
 module.exports = {
     data: new SlashCommandBuilder()
